Revert "base: official xposed framework commits by rovo89"
This reverts commit a1143cdaa11b5a641f0eeecb913959a0b1f5599a.

+++ cmds/app_process/xposed.cpp
@@ -16,7 +16,6 @@
 #include <dlfcn.h>
 #include <errno.h>
 #include <fcntl.h>
-#include <inttypes.h>
 #include <stdlib.h>
 
 #if PLATFORM_SDK_VERSION >= 18
@@ -313,7 +312,7 @@ static bool addPathToEnv(const char* name, const char* path) {
         char newPath[4096];
         int neededLength = snprintf(newPath, sizeof(newPath), "%s:%s", path, oldPath);
         if (neededLength >= (int)sizeof(newPath)) {
-            ALOGE("ERROR: %s would exceed %" PRIuPTR " characters", name, sizeof(newPath));
+            ALOGE("ERROR: %s would exceed %d characters", name, (int) sizeof(newPath));
             return false;
         }
         setenv(name, newPath, 1);
